pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ml-experiment-env"
version = "1.0.0"
description = "ML Experiment Reviewer - An OpenEnv environment for training AI agents to review ML experiments"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"
